% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepDat.R
\name{prepDat}
\alias{prepDat}
\title{prepare the data for analysis}
\usage{
prepDat(data,bma,gridsize)
}
\arguments{
\item{data}{list of ppp objects. Each entry of the list is a ppp object corresponding to a slice. Instead of a ppp object, it can be a named list with the following entries: x(numeric vector of x-coordinates of the locations of the observations), y(numeric vector of y-coordinates of the locations of the observations), and  marks(character vector of type or identifier of the observations that classifies them into the different possible species).}

\item{bma}{positive scalar. Thickness of each slice}

\item{gridsize}{numeric vector of size two with the entries being positive integers themselves. The desired dimensions of the computing grid.}
}
\value{
list of data products needed for input to Slices_to_3D. Object of class X
}
\description{
Prepares the data from different slices into the input format for the analysis function Slices_to_3D
}
