% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice_to_3D_ell.R
\name{Slice_to_3D_ell}
\alias{Slice_to_3D_ell}
\title{Draw posterior samples from the three dimensional process using two dimensional slices}
\usage{
Slice_to_3D_ell(data,K,
inits=NULL,
hyperparams=list("mu0" = rep(0,dim(data$Y)[2]), "rho0" = rep(1,K),
"A0" = matrix(0,dim(data$Y)[2],K),
"musig2"=0,"sigsig2"=10,
"sig2mu"=100,"sig2rho"=100,"sig2A"=100),
MCMCparams=list(niters=1.2e4,nthin=10,nburn=2e3,adapt=TRUE))
}
\arguments{
\item{data}{X class object or a list containing the entries array Y (number of slices x number of species x number of points), matrix S (number of locations x 2), vector gridsize (2x1, size of the computation grid) and positive scalar bma (thickness of each slice)}

\item{K}{positive integer, the number of latent processes to be used. Recommended to be smaller than the number of species}

\item{inits}{named list containing initial values for the parameters of the model. The components are mu(numeric vector, same length as the number of different species: 3 dimensional constant mean for each species), rho(positive numeric vector, same length as K: 3 dimensional spatial range parameter for each latent process), A(numeric matrix of dimension number of species x K: mapping matrix from the latent processes to the different species), sig2(positive numeric vector, same length as the number of slices: slice specific variability). If NULL (default), a quick initial value is selected within the function.}

\item{hyperparams}{named list containing the hyperparameters for the prior distributions of mu, A, rho, log(sig2). The components are mu0(numeric vector, same length as the number of different species: mean for the normal prior for the mu parameters), rho0(positive numeric vector, same length as K: exponentiated mean for the normal prior for the log(rho) parameters), A0(numeric matrix of dimension number of species x K: mean for the normal prior of the A parameters), musig2(scalar: mean parameter for the normal prior for log sig2), sigsig2(positive scalar: variance parameter for the normal prior for log sig2), sig2mu(positive scalar: variance parameter for the normal prior of the mu parameters), sig2A(positive scalar: variance parameter for the normal prior of the A parameters), sig2rho(positive scalar: variance parameter for the normal prior of the rho parameters).}

\item{MCMCparams}{named list containing details about the running of MCMC, the components are niters (integer, number of posterior samples to be drawn), nthin(integer, the thinning interval), nburn (integer, number of posterior samples to be discarded as burn-in samples), adapt (Boolean, whether an adaptive scheme should be used to make the acceptance ratio close to 0.574, defaults to TRUE). Total number of iterations = niters*nthin}
}
\value{
list of posterior samples for the parameters
}
\description{
Generates posterior samples for parameters of the three dimensional process using two dimensional slices
}
